Security related code updates.

+++ MVC/Controllers/DirectorsController.cs
@@ -36,26 +36,49 @@ namespace MVC.Controllers
             return View(list);
         }
 
+        bool IsOwnAccount(int id)
+        {
+            return id.ToString() == (User.Claims.SingleOrDefault(claim => claim.Type == "Id")?.Value ?? string.Empty);
+        }
+
         // GET: Directors/Details/5
         public IActionResult Details(int id)
         {
+            if (!IsOwnAccount(id) && !User.IsInRole("Admin"))
+            {
+                SetTempData("You are not authorized for this operation!");
+                return RedirectToAction(nameof(Index));
+            }
+
             var item = _directorService.Item(id);
             return View(item);
         }
 
         // GET: Directors/Create
-        [Authorize(Roles = "Admin")]
+        [Authorize]
         public IActionResult Create()
         {
+            if (!User.IsInRole("Admin"))
+            {
+                SetTempData("You are not authorized for this operation!");
+                return RedirectToAction(nameof(Index));
+            }
+
             SetViewData();
             return View();
         }
 
         // POST: Directors/Create
         [HttpPost, ValidateAntiForgeryToken]
-        [Authorize(Roles = "Admin")]
+        [Authorize]
         public IActionResult Create(DirectorRequest director)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                SetTempData("You are not authorized for this operation!");
+                return RedirectToAction(nameof(Index));
+            }
+
             if (ModelState.IsValid)
             {
                 var response = _directorService.Create(director);
@@ -71,9 +94,15 @@ namespace MVC.Controllers
         }
 
         // GET: Directors/Edit/5
-        [Authorize(Roles = "Admin")]
+        [Authorize]
         public IActionResult Edit(int id)
         {
+            if (!IsOwnAccount(id) && !User.IsInRole("Admin"))
+            {
+                SetTempData("You are not authorized for this operation!");
+                return RedirectToAction(nameof(Index));
+            }
+
             var item = _directorService.Edit(id);
             SetViewData();
             return View(item);
@@ -81,9 +110,15 @@ namespace MVC.Controllers
 
         // POST: Directors/Edit
         [HttpPost, ValidateAntiForgeryToken]
-        [Authorize(Roles = "Admin")]
+        [Authorize]
         public IActionResult Edit(DirectorRequest director)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                SetTempData("You are not authorized for this operation!");
+                return RedirectToAction(nameof(Index));
+            }
+
             if (ModelState.IsValid)
             {
                 var response = _directorService.Update(director);
@@ -99,18 +134,30 @@ namespace MVC.Controllers
         }
 
         // GET: Directors/Delete/5
-        [Authorize(Roles = "Admin")]
+        [Authorize]
         public IActionResult Delete(int id)
         {
+            if (!IsOwnAccount(id) && !User.IsInRole("Admin"))
+            {
+                SetTempData("You are not authorized for this operation!");
+                return RedirectToAction(nameof(Index));
+            }
+
             var item = _directorService.Item(id);
             return View(item);
         }
 
         // POST: Directors/Delete
         [HttpPost, ValidateAntiForgeryToken, ActionName("Delete")]
-        [Authorize(Roles = "Admin")]
+        [Authorize]
         public IActionResult DeleteConfirmed(int id)
         {
+            if (!IsOwnAccount(id) && !User.IsInRole("Admin"))
+            {
+                SetTempData("You are not authorized for this operation!");
+                return RedirectToAction(nameof(Index));
+            }
+
             var response = _directorService.Delete(id);
             SetTempData(response.Message);
             return RedirectToAction(nameof(Index));
